Ensure a correct list of tokens in http_tools.
This is a compatibility breaking change in the protocol. Tokens and requests are now numbered to make sure that no MITM attack can add or remove tokens in the body of a request or a response.
To also protect the request query, a hash or the PATH_INFO part or the query is passed in the initial token of a request.